requirements.txt
streamlit
opencv-python-headless
numpy
pillow
streamlit-image-coordinates
pandas